feat: move client redirect feature into valaxy
