+++ demo/yun/package.json
@@ -18,7 +18,6 @@
     "valaxy": "workspace:*",
     "valaxy-addon-algolia": "workspace:*",
     "valaxy-addon-bangumi": "workspace:*",
-    "valaxy-addon-client-redirects": "workspace:*",
     "valaxy-addon-components": "workspace:*",
     "valaxy-addon-lightgallery": "workspace:*",
     "valaxy-addon-meting": "workspace:*",
