--- packages/valaxy-addon-client-redirects/index.d.ts
@@ -1,2 +1,0 @@
-export * from './types'
-export * from './node'
