--- packages/valaxy-addon-client-redirects/index.ts
@@ -1,1 +1,0 @@
-export * from './node'
