--- packages/valaxy-addon-client-redirects/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "valaxy-addon-client-redirects",
-  "version": "0.0.1",
-  "description": "Client Redirect for Valaxy",
-  "repository": "https://github.com/YunYouJun/valaxy/tree/main/packages/valaxy-addon-client-redirects",
-  "keywords": [
-    "valaxy",
-    "addon",
-    "client-redirects"
-  ],
-  "main": "index.ts",
-  "types": "index.d.ts",
-  "peerDependencies": {
-    "valaxy": "latest"
-  },
-  "dependencies": {
-    "fs-extra": "^11.2.0"
-  }
-}
