--- packages/valaxy-addon-client-redirects/types/index.ts
@@ -1,8 +1,0 @@
-export interface RedirectRule {
-  to: string
-  from: string | string[]
-}
-
-export interface ClientRedirectsOptions {
-  redirects: RedirectRule[]
-}
