+++ packages/valaxy/node/utils/clientRedirects.ts
@@ -1,4 +1,6 @@
-import type { RedirectRule } from '../types'
+import { writeFile } from 'node:fs/promises'
+import { ensureFile } from 'fs-extra'
+import type { RedirectRule } from '../../types'
 
 function handleRoute(route: string) {
   if (route === '/')
@@ -33,4 +35,21 @@ export function collectRedirects(redirectRule: RedirectRule[]) {
   }
 
   return redirects
+}
+
+export async function writeRedirectFiles(route: string, filePath: string) {
+  await ensureFile(filePath)
+  await writeFile(filePath, `
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="refresh" content="0; url=${route}">
+  <link rel="canonical" href="${route}">
+</head>
+  <script>
+    window.location.href = '${route}' + window.location.search + window.location.hash
+  </script>
+</html>
+  `)
 }
